# "Loop labels" greedy decoding: faster implementation