--- a/nemo/collections/asr/modules/rnnt.py
+++ b/nemo/collections/asr/modules/rnnt.py
@@ -368,6 +368,27 @@ class StatelessTransducerDecoder(rnnt_abstract.AbstractRNNTDecoder, Exportable):
 
         return state_list
 
+    def batch_replace_states(
+        self,
+        src_states: Optional[List[torch.Tensor]],
+        src_mask_or_indices: torch.Tensor,
+        dst_states: Optional[List[torch.Tensor]],
+        dst_mask_or_indices: torch.Tensor,
+    ):
+        if src_states is None or dst_states is None:
+            # NB: it is important for torch.jit that the state has an optional type consistent with other code
+            # So, we have to check this and always annotate with Optional[List[torch.Tensor]]
+            return  # nothing to replace
+        dst_states[0][dst_mask_or_indices] = src_states[0][src_mask_or_indices]
+
+    def batch_replace_states_mask(
+        self, src_states: list[torch.Tensor], dst_states: list[torch.Tensor], mask: torch.Tensor,
+    ):
+        torch.where(mask.unsqueeze(-1), src_states[0], dst_states[0], out=dst_states[0])
+
+    def batch_split_states(self, batch_states: list[torch.Tensor]) -> list[list[torch.Tensor]]:
+        return [sub_state.split(1, dim=0) for sub_state in batch_states]
+
     def batch_copy_states(
         self,
         old_states: List[torch.Tensor],
@@ -790,7 +811,7 @@ class RNNTDecoder(rnnt_abstract.AbstractRNNTDecoder, Exportable, AdapterModuleMi
         )
         return layers
 
-    def initialize_state(self, y: torch.Tensor) -> List[torch.Tensor]:
+    def initialize_state(self, y: torch.Tensor) -> Tuple[torch.Tensor, torch.Tensor]:
         """
         Initialize the state of the RNN layers, with same dtype and device as input `y`.
 
@@ -805,16 +826,16 @@ class RNNTDecoder(rnnt_abstract.AbstractRNNTDecoder, Exportable, AdapterModuleMi
         """
         batch = y.size(0)
         if self.random_state_sampling and self.training:
-            state = [
+            state = (
                 torch.randn(self.pred_rnn_layers, batch, self.pred_hidden, dtype=y.dtype, device=y.device),
                 torch.randn(self.pred_rnn_layers, batch, self.pred_hidden, dtype=y.dtype, device=y.device),
-            ]
+            )
 
         else:
-            state = [
+            state = (
                 torch.zeros(self.pred_rnn_layers, batch, self.pred_hidden, dtype=y.dtype, device=y.device),
                 torch.zeros(self.pred_rnn_layers, batch, self.pred_hidden, dtype=y.dtype, device=y.device),
-            ]
+            )
         return state
 
     def score_hypothesis(
@@ -1029,6 +1050,35 @@ class RNNTDecoder(rnnt_abstract.AbstractRNNTDecoder, Exportable, AdapterModuleMi
             state_list.append(state_tensor)
 
         return state_list
+
+    def batch_replace_states(
+        self,
+        src_states: Tuple[torch.Tensor, torch.Tensor],
+        src_mask_or_indices: torch.Tensor,
+        dst_states: Tuple[torch.Tensor, torch.Tensor],
+        dst_mask_or_indices: torch.Tensor,
+    ):
+        # TODO: why do we need to cast?
+        #  with bfloat16 the initialize_state returns bfloat16, but lstm returns float16
+        dst_states[0][:, dst_mask_or_indices] = src_states[0][:, src_mask_or_indices].to(dst_states[0].dtype)
+        dst_states[1][:, dst_mask_or_indices] = src_states[1][:, src_mask_or_indices].to(dst_states[1].dtype)
+
+    def batch_replace_states_mask(
+        self,
+        src_states: Tuple[torch.Tensor, torch.Tensor],
+        dst_states: Tuple[torch.Tensor, torch.Tensor],
+        mask: torch.Tensor,
+    ):
+        # TODO: why do we need to cast?
+        #  with bfloat16 the initialize_state returns bfloat16, but lstm returns float16
+        dtype = dst_states[0].dtype
+        torch.where(mask.unsqueeze(0).unsqueeze(-1), src_states[0].to(dtype), dst_states[0], out=dst_states[0])
+        torch.where(mask.unsqueeze(0).unsqueeze(-1), src_states[1].to(dtype), dst_states[1], out=dst_states[1])
+
+    def batch_split_states(
+        self, batch_states: Tuple[torch.Tensor, torch.Tensor]
+    ) -> list[Tuple[torch.Tensor, torch.Tensor]]:
+        return list(zip(batch_states[0].split(1, dim=1), batch_states[1].split(1, dim=1)))
 
     def batch_copy_states(
         self,