… scripts/tts_dataset_files/ipa_cmudict-0.7b_nv23.01.txt